+++ src/GrahamCampbell/CoreAPI/Classes/CoreAPI.php
@@ -1,8 +1,27 @@
 <?php namespace GrahamCampbell\CoreAPI\Classes;
 
-use Illuminate\Support\Facades\Cache;
-use Illuminate\Support\Facades\Config;
+/**
+ * This file is part of Laravel Core API by Graham Campbell.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ * @package    Laravel-Core-API
+ * @author     Graham Campbell
+ * @license    Apache License
+ * @copyright  Copyright 2013 Graham Campbell
+ * @link       https://github.com/GrahamCampbell/Laravel-Core-API
+ */
 
+use Illuminate\Cache\CacheManager;
+use Illuminate\Config\Repository;
 use Guzzle\Common\Collection;
 use Guzzle\Http\Client;
 use Guzzle\Plugin\Oauth\OauthPlugin;
@@ -10,16 +29,35 @@ use Guzzle\Plugin\Oauth\OauthPlugin;
 class CoreAPI {
 
     protected $baseurl;
-    protected $config;
+    protected $conf;
     protected $oauth;
     protected $auth;
     protected $userAgent;
     protected $client;
 
-    public function setUp($baseurl, $config = array(), $authentication = null, $userAgent = null) {
+    /**
+     * The cache instance.
+     *
+     * @var \Illuminate\Cache\CacheManager
+     */
+    protected $cache;
+
+    /**
+     * The config instance.
+     *
+     * @var \Illuminate\Config\Repository
+     */
+    protected $config;
+
+    public function __construct(CacheManager $cache, Repository $config) {
+        $this->cache = $cache;
+        $this->config = $config;
+    }
+
+    public function setUp($baseurl, $conf = array(), $authentication = null, $userAgent = null) {
         $this->baseurl = $baseurl;
 
-        $this->config = new Collection($config);
+        $this->conf = new Collection($conf);
 
         if (isset($authentication['user'])) {
             $this->auth = $authentication;
@@ -35,7 +73,7 @@ class CoreAPI {
     }
 
     protected function makeNewClient() {
-        $this->client = new Client($this->baseurl, $this->config);
+        $this->client = new Client($this->baseurl, $this->conf);
 
         if ($this->oauth) {
             $this->setOauth($this->oauth);
@@ -66,9 +104,9 @@ class CoreAPI {
         return $this->client->getConfig();
     }
 
-    public function setConfig($config) {
-        $this->config = new Collection($config);
-        return $this->client->setConfig($this->config);
+    public function setConfig($conf) {
+        $this->conf = new Collection($conf);
+        return $this->client->setConfig($this->conf);
     }
 
     public function getOauth() {
@@ -126,32 +164,33 @@ class CoreAPI {
     }
 
 
-    public function getAllEvents() {
-        return $this->client->getAllEvents();
-    }
+    public function goGet($method = 'GET', $uri = null, $headers = null, $body = null, array $options = array(), $cache = false) {
+        $key = md5(json_encode($method).json_encode($uri).json_encode($headers).json_encode($body).json_encode($options));
+        $time = $this->cacheTime($cache);
 
-    public function createRequest($method = 'GET', $uri = null, $headers = null, $body = null, array $options = array()) {
-        return $this->client->createRequest($method, $uri, $headers, $body, $options);
-    }
-
-    public function send($requests) {
-        return $this->client->send($requests);
-    }
-
-
-    protected function cache($values, $func, $type, $cache) {
-        $cache = $this->shouldCache($cache);
-
-        if ($cache !== 0) {
-            return $this->pullCache($values, $func, $type, $cache);
+        // check if we are using the cache
+        if ($time > 0) {
+            // if so, then pull from the cache
+            $value = $this->getCache($key);
+            // check if the value is valid
+            if (!$this->validCache($value)) {
+                // if is invalid, do the work
+                $value = $this->sendGet($method, $uri, $headers, $body, $options);
+                // add the value from the work to the cache
+                $this->setCache($key, $value, $time);
+            }
+        } else {
+            // do the work because caching is disabled
+            $value = $this->sendGet($method, $uri, $headers, $body, $options);
         }
 
-        return $this->getBody($values, $func, $type);
+        // spit out the response
+        return $value;
     }
 
-    protected function shouldCache($cache) {
+    protected function cacheTime($cache) {
         if ($cache === true) {
-            $cache = Config::get('core-api::cache');
+            $cache = $this->config['core-api::cache'];
         } elseif ($cache === false) {
             $cache = 0;
         } elseif (is_numeric($cache)) {
@@ -162,157 +201,58 @@ class CoreAPI {
             $cache = 0;
         }
 
-        if (Config::get('core-api::cache') === 0 && Config::get('core-api::force') === true) {
+        if ($cache  < 0) {
+            $cache = 0;
+        }
+
+        if ($this->config['core-api::cache'] === 0 && $this->config['core-api::force'] === true) {
             $cache = 0;
         }
 
         return $cache;
     }
 
-    protected function pullCache($values, $func, $type, $cache) {
-        $key = $this->getKey($values, $type);
+    protected function getCache($key) {
+        return $this->cache->section('api')->get($key);
+    }
 
-        if (Cache::section('api')->has($key)) {
-            return Cache::section('api')->get($key);
+    protected function validCache($value) {
+        if (is_null($value) || !is_a($value, '\GrahamCampbell\CoreAPI\Classes\APIResponse')) {
+            return false;
         }
 
-        $value = $this->getBody($values, $func, $type);
-
-        Cache::section('api')->put($key, $value, $cache);
-
-        return $value;
+        return true;
     }
 
-    protected function getKey($values, $type) {
-        switch ($type) {
-            case 'get':
-            case 'head':
-                return md5(json_encode($type).json_encode($values['uri']).json_encode($values['headers']).json_encode($values['options']));
-            case 'options':
-                return md5(json_encode($type).json_encode($values['uri']).json_encode($values['options']));
-            default:
-                return md5(json_encode($type).json_encode($values['uri']).json_encode($values['headers']).json_encode($values['body']).json_encode($values['options']));
-        }
+    protected function sendGet($method, $uri, $headers, $body, $options) {
+        $request = $this->client->createRequest($method, $uri, $headers, $body, $options);
+        return new APIResponse($request);
     }
 
-    protected function getBody($values, $func, $type) {
-        switch ($type) {
-            case 'get':
-            case 'head':
-                return json_decode($func($this->client, $values['uri'], $values['headers'], $values['options']), true);
-            case 'options':
-                return json_decode($func($this->client, $values['uri'], $values['options']), true);
-            default:
-                return json_decode($func($this->client, $values['uri'], $values['headers'], $values['body'], $values['options']), true);
-        }
+    protected function setCache($key, $value, $time) {
+        return $this->cache->section('api')->put($key, $value, $time);
     }
 
 
-    public function get($uri = null, $headers = null, $options = array()) {
-        return $this->client->get($uri, $headers, $options);
+    public function get($uri = null, $headers = null, $options = array(), $cache = false) {
+        return is_array($options)
+            ? $this->goGet('GET', $uri, $headers, null, $options, $cache)
+            : $this->goGet('GET', $uri, $headers, $options, array(), $cache);
     }
 
-    public function sendGet($uri = null, $headers = null, $options = array()) {
-        return $this->client->get($uri, $headers, $options)->send();
+    public function post($uri = null, $headers = null, $body = null, $options = array(), $cache = false) {
+        return $this->goGet('POST', $uri, $headers, $body, $options, $cache);
     }
 
-    public function goGet($uri = null, $headers = null, $options = array(), $cache = true) {
-        return $this->cache(array('uri' => $uri, 'headers' => $headers, 'options' => $options),
-            function($client, $uri, $headers, $options) {
-                if (isset($this->auth['user'])) {
-                    return $client->get($uri, $headers, $options)->setAuth($this->auth['user'], $this->auth['pass'])->send()->getBody();
-                }
-                return $client->get($uri, $headers, $options)->send()->getBody();
-        }, 'get', $cache);
+    public function put($uri = null, $headers = null, $body = null, $options = array(), $cache = false) {
+        return $this->goGet('PUT', $uri, $headers, $body, $options, $cache);
     }
 
-    public function head($uri = null, $headers = null, $options = array()) {
-        return $this->client->head($uri, $headers, $options);
+    public function patch($uri = null, $headers = null, $body = null, $options = array(), $cache = false) {
+        return $this->goGet('PATCH', $uri, $headers, $body, $options, $cache);
     }
 
-    public function sendHead($uri = null, $headers = null, $options = array()) {
-        return $this->client->head($uri, $headers, $options)->send();
-    }
-
-    public function goHead($uri = null, $headers = null, $options = array(), $cache = false) {
-       return $this->cache(array('uri' => $uri, 'headers' => $headers, 'options' => $options),
-            function($client, $uri, $headers, $options) {
-                return $client->head($uri, $headers, $options)->send()->getBody();
-        }, 'head', $cache);
-    }
-
-    public function delete($uri = null, $headers = null, $body = null, $options = array()) {
-        return $this->client->delete($uri, $headers, $body, $options);
-    }
-
-    public function sendDelete($uri = null, $headers = null, $body = null, $options = array()) {
-        return $this->client->delete($uri, $headers, $body, $options)->send();
-    }
-
-    public function goDelete($uri = null, $headers = null, $body = null, $options = array(), $cache = false) {
-        return $this->cache(array('uri' => $uri, 'headers' => $headers, 'body' => $body, 'options' => $options),
-            function($client, $uri, $headers, $body, $options) {
-                return $client->delete($uri, $headers, $body, $options)->send()->getBody();
-        }, 'delete', $cache);
-    }
-
-    public function put($uri = null, $headers = null, $body = null, $options = array()) {
-        return $this->client->put($uri, $headers, $body, $options);
-    }
-
-    public function sendPut($uri = null, $headers = null, $body = null, $options = array()) {
-        return $this->client->put($uri, $headers, $body, $options)->send();
-    }
-
-    public function goPut($uri = null, $headers = null, $body = null, $options = array(), $cache = false) {
-        return $this->cache(array('uri' => $uri, 'headers' => $headers, 'body' => $body, 'options' => $options),
-            function($client, $uri, $headers, $body, $options) {
-                return $client->put($uri, $headers, $body, $options)->send()->getBody();
-        }, 'put', $cache);
-    }
-
-    public function patch($uri = null, $headers = null, $body = null, $options = array()) {
-        return $this->client->patch($uri, $headers, $body, $options);
-    }
-
-    public function sendPatch($uri = null, $headers = null, $body = null, $options = array()) {
-        return $this->client->patch($uri, $headers, $body, $options)->send();
-    }
-
-    public function goPatch($uri = null, $headers = null, $body = null, $options = array(), $cache = false) {
-        return $this->cache(array('uri' => $uri, 'headers' => $headers, 'body' => $body, 'options' => $options),
-            function($client, $uri, $headers, $body, $options) {
-                return $client->patch($uri, $headers, $body, $options)->send()->getBody();
-        }, 'patch', $cache);
-    }
-
-    public function post($uri = null, $headers = null, $body = null, $options = array()) {
-        return $this->client->post($uri, $headers, $body, $options);
-    }
-
-    public function sendPost($uri = null, $headers = null, $body = null, $options = array()) {
-        return $this->client->post($uri, $headers, $body, $options)->send();
-    }
-
-    public function goPost($uri = null, $headers = null, $body = null, $options = array(), $cache = false) {
-        return $this->cache(array('uri' => $uri, 'headers' => $headers, 'body' => $body, 'options' => $options),
-            function($client, $uri, $headers, $body, $options) {
-                return $client->post($uri, $headers, $body, $options)->send()->getBody();
-        }, 'post', $cache);
-    }
-
-    public function options($uri = null, array $options = array()) {
-        return $this->client->options($uri, $options);
-    }
-
-    public function sendOptions($uri = null, array $options = array()) {
-        return $this->client->options($uri, $options)->send();
-    }
-
-    public function goOptions($uri = null, array $options = array(), $cache = false) {
-        return $this->cache(array('uri' => $uri, 'options' => $options),
-            function($client, $uri, $headers, $body, $options) {
-                return $client->options($uri, $options)->send()->getBody();
-        }, 'options', $cache);
+    public function delete($uri = null, $headers = null, $body = null, $options = array(), $cache = false) {
+        return $this->goGet('DELETE', $uri, $headers, $body, $options, $cache);
     }
 }